LayoutTests/fast/box-shadow/box-shadow-overflow-scroll-expected.txt

Check if the size of box-shadow's blur and spread are considered as overflow size. See this bug for detail: https://bugs.webkit.org/show_bug.cgi?id=28017.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.scrollLeft is 40+50+60
PASS element.scrollTop is 40+50+60
PASS element.scrollLeft is 40+50+60
PASS element.scrollTop is 40+50+60
PASS element.scrollLeft is 40+50+60
PASS element.scrollTop is 40+50+60
PASS successfullyParsed is true

TEST COMPLETE
